+++ artemis/modules/utils/wappalyzer/go.mod
@@ -2,6 +2,6 @@ module wappalyzer
 
 go 1.23.6
 
-require github.com/projectdiscovery/wappalyzergo v0.2.35
+require github.com/projectdiscovery/wappalyzergo v0.2.36
 
 require golang.org/x/net v0.41.0 // indirect
